% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMaxt4.R
\docType{package}
\name{AMaxt4}
\alias{AMaxt4}
\title{AMax-T test package}
\description{
This package find test statistic value critical value of
asymptotic Max-T test for testing
equality of means against ordered alternative in one-way ANOVA model
with heteroscedasticity when number of groups is 4.
}
\author{
Anjana Mondal \email{mondal.anjana25@gmail.com}
}
